Check for --apparent-size before invoking on cwd.
In general, it is possible for du to emit output and still return failure.

Previously, this resulted in duplicate entries on systems that support both long
and short options (since both parts of the AND-OR are run).

Here, we prevent duplicates by only having one outputting du invocation, at the
cost of extra assumptions:

- /dev/null is readable and length 0.
- Short options are always supported.