directory: Drop basename_is_rel_to logic
We can't set a `Directory` with multiple `basename_is_rel_to` values and
it can appear simultaneously in both flattened and non-flattened
directories, or differently flattened directories, so it can in fact be
simultaneously rendered relative to multiple directories.

The `relative_path` property of an `FSObject` is now always semantically
meaningful, relative to the current working directory. This does mean
`relative_path` is no longer usable as a way to get an `FSObject`'s
"display name."