linemode: Fix using appropriate "display names"
Changes to `Directory` mean the `relative_path` attribute is no longer a
(bad) proxy for the way the object is displayed in the browser columns.

This means line modes can no longer use it as such. Using the path
relative to the target of the current column, or the basename otherwise
seems like an appropriate alternative.