directory: Fix hidden filter after relative_path change
While working on issues related to flattening I noticed unfortunate
interplay between being in a hidden directory and the hidden filter.
If we are in a hidden directory there's no point in hiding that
directory from the parent directory. It causes confusion because the
directory we are in does not appear in the parent and instead a
different child of the parent is highlighted, leading us to potentially
believe we are in a different directory.

I've also stopped filtering away hidden files and directories if they
are highlighted, which makes moving left (to the parent) and then right
(into the highlighted child) again a more idempotent operation.