FM: don't modify UI while rifle is running
When Rifle is running a child process, such as Vim, and the user presses
Ctrl-Z, both Vim and Ranger receive the signal to suspend. This is
because they are both in the terminal's "foreground process group"
according to POSIX documentation:

> If the ISIG flag is set, receipt of the [Ctrl-Z] character shall cause a SIGTSTP
> signal to be sent to all processes in the foreground process group for which
> the terminal is the controlling terminal ...
https://archive.ph/YSNB3

Since both Vim and Ranger have signal handlers, both of their signal
handlers will execute when their group receives a signal. This leads to problems
when it comes to the signals SIGTSTP (suspend) and SIGCONT (resume).
The problem is that when Vim and Ranger receives SIGCONT/SIGTSTP,
they both try to modify the terminal settings simultaneously, which is a race condition.
In some situations the result can be a mixture of both terminal
settings, which messes up Vim's UI even while it is still running.

To avoid this, Ranger should not try to modify the UI while Rifle is
running some other process in the foreground.