Remove duplicate tabs on startup
With the `save_tabs_on_exit=true` option, every time Ranger is started
without explicitly specifying which tabs to open, a new tab is first
created for the current working directory, and then the previously saved
tabs from the last session are added. So if the current directory was
already in the list from the previous session, then it was saved twice
on the next exit.
Try opening multiple tabs in Ranger, and then opening and closing Ranger
from the same path multiple times in a row. You will see that every time
you run it, a new tab will be created with the same path. Since it
doesn't make much sense to keep multiple tabs with the same path,
I suggest removing duplicates from `fm.start_paths` at startup.