Don't strip commands of trailing spaces
It is nice be able to have trailing spaces in commands because it reduces the
amount of keystrokes the user needs to use. For example if a user wants
to create a shortcut for `touch` they'd like to have a trailing space
which allows them to immediatly start typing the name of the file they
are creating.