scope.sh: Enable DjVu previews by default
ImageMagick requires a large amount of memory to process DjVu files and
is run on anything with a MIME type matching `image/*`. By enabling the
special case for DjVu files by default, we either use `ddjvu` or we exit
instead of trying to run them through ImageMagick. Preventing this from
happening.

PR #1947 tried to address this similarly. The difference is it uses
ImageMagick's convert on the ddjvu output to make sure it produces a
JPEG. This shouldn't really be necessary and avoiding ImageMagick
altogether seems nicer in this case.

Fixes #2366
Closes #1947