ranger.py: Allow passing -flags and --options when sourcing
A previous refactor prefixed any arguments passed when sourcing the
`ranger.py` polyglot with ` -- ` in an attempt to prevent accidentally
interpreting a path as flags or options. Because of the quoting involved
this is not actually necessary and only serves to prevent passing flags
and options when sourcing the script.

Another thing that came up in discussion for PR #1743 is whether or not
to explicitly replace an empty argument list with the PWD. In particular
because of concerns around relying on Ranger's implicit behavior of
opening the PWD when no arguments are passed. Since this behavior is
already relied upon when any arguments that are not paths are passed, it
should be reasonable to rely on this behavior when no arguments are
passed. Furthermore, if the implicit behavior does ever change it will
prevent a possibly unexpected difference in behavior.

Fixes #2187