img_display: Restore cleanup of Kitty tempfiles
Kitty version 0.26.0 stopped removing temporary files used in the Kitty
Image Protocol unless their full path contains the string
"tty-graphics-protocol," so we need to adapt to the new behavior.

By using a subdirectory of the directory used for temporary files with
the appropriate name, Kitty should again delete the temporary PNGs we
generate. An added benefit of using a subdirectory is avoiding
cluttering up the directory for temporary files. This subdirectory is
not cleaned up by Ranger and will remain as an empty directory (if
previews work correctly) until removed by other means.

Fixes #2699