directory: Promote is_file assertion to exception
Assertions are only checked when Python is run without the optimize flag
`-O`, which we have in our default shebang.

Certain situations can lead to a path to a file being used to
instantiate a `Directory` object and we need to be able to check for
this exceptional condition even if Python's run with optimization
enabled.