bookmarks: Catch and report erroneous path exception
If a path is saved as a bookmark, which can be automatic for the last
visited mark `'` for example, and the directory is replaced by a file
then Ranger will crash when trying to instantiate a `Directory` object
for that path.

We catch this exception and report it to the user. Upon successful
saving of the bookmarks upon quitting Ranger, such paths will not be
saved again, essentially removing them.

Fixes #1896
Fixes #2007