scope: Resolve unreachable exit
The `exit 1` at the end of handle_fallback makes the `exit 1` at the end
of the file unreachable. Shellcheck recommends eliminating unreachable
code. We have two options here and opted to drop the exit at the end of
handle_fallback which means all the functions are semantically equal,
they exit in case of a match, fall through in case of no match.