actions: Replace SystemExit with RuntimeError
SystemExit is treated specially by the Python interpreter, it silences
the usual behavior of printing a traceback for exceptions.

RuntimeError is not quite appropriate as this is not an error, ending
with a stacktrace is the desired result. But other built-in exceptions
aren't more appropriate. A custom exception might be a nice indication
but we can refactor again if we decide it's desireable.