img_display: Directory existing is not an error
Caught in review by socksthefennec, os.mkdir issues a FileExistsError if
the path already exists. I'm ignoring the possibility that the path
exists and points to a file for now.

For Python 2.7 an OSError is raised instead so we need to distinguish
that error in the except clause.

An alternative solution is to use os.makedirs, which is more like `mkdir
-p` and has an `exists_ok` parameter, unfortunately that parameter was
not present in Python 2.7.

Fixes #2775