gui.color: Change dummy key
I believe "10" came from foreground white "1" and background black "0"
but the keys to this dictionary are pairs, not concatenated strings. So
it seems more accurate to put the unchangeable pair in the dictionary.
Additionally, we don't want to waste a precious color pair on the
white/black pair because that's always wired in as pair 0.