color: Catch Color number is greater than COLORS-1
Some terminals do not support the number of color pairs we use. Python's
curses module used to throw a curses.error for this case but in Python
3.10 this was changed to ValueError.

Approximating the pair with the closest one that is already set up might
be a viable fix but very few terminals are so limited in their color
support and it might make things subtly off but hard to put your finger
on. So we are falling back to the default white on black pair.

Fixes #2498
Closes #2741
Fixes #2752
Fixes #2761
Fixes #2800