loader: Accumulate stderr for commands
We accumulate stdout already because we're reading it in chunks. We
should do the same for stderr as well, otherwise we can get long
sequences of partial `fm.notify`s.

Note that reads of both stdout and stderr can be partial. We read up to
4096 bytes at a time until the process has finished. We may not have
read all of the process's output at that time.