This is a combination of 6 commits
This is the 1st commit message:

Close with confirmation the "dead" tab when switching through it

This is the commit message #2:

Prevent another tab_move call before answering the current question

This is the commit message #3:

Close dead tabs with confirmation when switching through them

Add a boolean setting `confirm_to_close_dead_tab` to specify the
behavior when switching tabs through a tab with a "dead" path and delete
that tab with confirmation or skip it silently otherwise.

This is the commit message #4:

Manage mouse clicks on tabs with "dead" path

This is the commit message #5:

Respect "confirm_to_close_dead_tab" setting for mouse events on dead tabs

This is the commit message #6:

Move confirmation of closing a dead tab to a separate function

To reuse it in another module.