Fix the ABC checks on Python 3
The previous solution used the ABCMeta metaclass directly as the ABC
non-meta class didn't exist in Python 2.

Python 2 and 3 have different syntax for metaclasses.  The one used by
Python 2 doesn't work on Python 3 but is not an error either.  The one
used by Python 3 is an error in Python 2.  We used to use the Python 2
syntax while acknowledging it merely doesn't do anything on Python 3.

The regular class ABC allows to avoid using the metaclass syntax at
all, but it's only available in Python 3.  This solution implements it
on our side (using the Python 2 metaclass syntax) if it's not
available, but otherwise uses the standard ABC class with
regular inheritance.