get_executables: Use platform instead of reading /proc
Not every system has a `/proc` file system. We could catch the error but
it is simpler to check `platform.release()`. This string contains `WSL2`
for WSL version 2 but the only common substring for version 1 is lower
case `microsoft`. On Windows this string should simply contain a major
version number and not include that substring.

Fixes #2815