directory: Fix max call in mtimelevel
Pylint flagged a nested max call as redundant but it was a false
positive. However, the nested max call materialized a generator as a
list and prepended `-1` because the function errors on an empty
argument. Instead of nesting a max call we can drop the nesting and
prepend the current value instead of passing it as a separate argument.

Fixes #2812