GHActions: Change CI to run conditionally without remaining pending
Using the built-in paths filter means Actions will not be run at all.
When they are set as "required checks" on PRs that means they will be
marked as pending forever. This is confusing for contributors and means
maintainers need to keep in mind when checks should actually complete or
not.

Using the paths-filter Action means we can use conditional execution of
steps, which results in things being marked as skipped, rather than
pending. This should make everything a bit less confusing.

Some tests that are only run on push events, mostly for the master
branch, still use the built-in path filtering. These aren't used as
required checks so they aren't as front and center.

One unfortunate consequence of the change is that more Action runs will
be spun up unnecessarily. Such senseless waste of resources is pretty
sad and GitHub should fix this.