ui: Refresh window in initialize
NCurses errors when `endwin()` is called twice in a row without the
window being refreshed.

Runner suspends the window (calling `endwin()` indirectly) and uses
`initialize` to resume the curses window. The latter was originally
envisioned as more of a one-item setup and only refreshes the window the
first time it is called. Refreshing unconditionally avoids the problem.

Fixes #2934
Closes #2935
Fixes #3000
Fixes #3005