GHActions: Use Pypy 3.10
This is currently the latest available Pypy.

Sadly it is still not possible to specify "pypy3.x" with the
"setup-python" action, actions/setup-python#646. While specifying
"python3.x" is considered useful, and "pypy3.8-v7.x" is considered
useful, that same logic does not seem to carry over to wanting whatever
is the latest Pypy regardless of what version of CPython it is
compatible with.