GHActions: Pypy don't run old Flake8/Pylint
Pypy 2.7 is stuck with old versions of Flake8 and Pylint.
Python 2 era Flake8 errors on invalid (Python 3) syntax.
Pylint has changed sufficiently to make it impossible to satisfy both
new and old versions.

As a workaround, only run Flake8 if its version is 7.1.1 or later and
only run Pylint if its version is 3.3.1 or later.

The versions available for Pypy 3 will probably have recent enough
Flake8 and Pylint to be compatible with the versions we run in the CI
action for Python 3. If this assumption fails the bounds may have to be
bumped manually.