img_display: Add support for iTerm2's v3.5.0+ image protocol
iTerm2 v3.5.0 introduced a change which imposes a hard limit of 1 MiB on
the length of escape sequences it will parse. Escape sequences longer
than this will force iTerm2 to bail on parsing, which results in the
file content spilling into the terminal as visible base64 characters.

To retain support for large files, it introduced a new 'multipart'
protocol variant intended to transfer a file as a series of escape
sequences rather than a single big one. This change updates
`ITerm2ImageDisplayer` to use this newer variant.

Also added is a new setting, `iterm2_multipart_file_chunk_size`, which
determines the chunk size. In many (hopefully most) cases, the default
added to rc.conf should suffice.  I even tried to avoid adding it, but
reference implementations and documentation suggest there are a handful
of situations where a user may want to tweak this value to make image
display work, and I don't believe these situations can be reliably
detected automatically. "Old" versions of tmux get a specific mention,
and anyone using iTerm versions older than 3.5.0 will want to force the
old protocol, which can be done with a value of 0.