add assertion because we assume toggle_ui is false for forked processes
If toggle_ui were True when we spawn a forked process,
we would have to make sure to track them in Runner's ProcessSet
until they exit, exactly like we do with non-forked processes.
Tracking a forked process will require some refactoring because it's not
as straightforward.

We have to track all `toggle_ui` processes in the ProcessSet, because `FM` will
check this ProcessSet to know when it's allowed to modify the terminal
mode. `FM` may only modify the terminal mode when ProcessSet's ui set is
empty (no UI processes are running).