Revert to original design of `delay_process_signals`
Signals such as `SIGTSTP` and `SIGCONT` may be raised by the user at
any time, including while the function is doing its bookkeeping. And
such signals are handled by both Ranger and its subprocesses. Depending
on the timing of the signals, Ranger may end up suspended while its
subprocesses are not, or vice versa. Because we can't control the timing
of the signals or make sure our subprocesses are suspended only when we
are, `_delay_process_signals` is inherently racy.

The previous fix (commit 1220a61cf4f769f32532b44ce3268914da7c9310)
also didn't truly address the underlying problem, so we lose little by
reverting to the original simpler version.