gui.ui: make suspend() more robust against flukes
During debugging (f.e. with rich.inspect or objexplore.explore), it
is often useful to call fm.ui.suspend() in order to temporarilty
give non-ranger code control over the terminal. However, there was
a change in ncurses on 20231111 that causes an error to be thrown
when endwin() is called multiple times without doupdate().
Just ignore it™.