Make ":open_with" command selection more precise (#3087)
When `:open_with` is used with a number, that number will
now always correspond to the same command as was on the list
shown to the user.

Also if `:open_with` is used when a list of commands isn't visible,
Ranger won't use the number to select a commands.

Rifle's behavior when run standalone should remain the same
after this commit.