CopyLoader: fix percentage calculation
The percentage bar was always calculated in terms of
number of bytes, not number of units `step`.
Mixing the units causes a visual bug where the percentage
bar doesn't reflect actual progress.

This fix also improves performance since we don't traverse
the tree twice to calculate the total size in different units.