GHActions: Pypy drop linting
Pypy 2.7 is stuck with older versions of Flake8 and Pylint. Old Pylint
would be attractive to run the Python 2 compatibility checks that have
since been removed but we can't run all the lints because old and new
Pylint can disagree, which would make it impossible to satisfy both CI
checks.

Pypy 3 seems to lag behind CPython at least on GHActions, this leads to
version mismatches, at the time of this commit Flake8 is at version
7.4.0 for CPython and 7.3.0 for Pypy 3.10 for example. This too can make
it impossible to satisfy both CI checks unless we constrain the versions
of the linters.

We unconstrained the linters intentionally to keep up with changes
gradually rather than having to do sweeping refactors every time we
decide to bump the linter versions.

Functional tests are more important than linting so we'll have to rely
on those to catch syntactical regressions. This isn't ideal because our
tests don't have great coverage but it's the most practical way forward.